Change handling of small files
When in "no overwrite"/"recovery" mode, treat files less than 70 bytes
as if they were not already present so that a new fetch is done in
attempt to get better replacement.
Token expiry when trying to get files (certainly from yimg.com) lead to
failed downloads that created 0 byte files - so change handling to get
new token if fileinfo.json is over 45 minutes old (token seen to expirty
around 60 minutes) PLUS do not create the output File if the download
failed so that 0 byte files not created.